(chore) add datatypes from emrapi and corresponding mocks
