+++ packages/esm-ward-app/src/ward-view/ward-bed.component.tsx
@@ -1,10 +1,10 @@
 import { type Visit, type Patient } from '@openmrs/esm-framework';
 import React from 'react';
 import EmptyBed from '../beds/empty-bed.component';
-import { type Bed } from '../types';
+import { type WardPatient, type Bed } from '../types';
 import OccupiedBed from '../beds/occupied-bed.component';
 export interface WardBedProps {
-  patientInfos: Array<{ patient: Patient; visit: Visit }>;
+  patientInfos: Array<WardPatient>;
   bed: Bed;
 }
